Timestamps are now correctly stored as numbers and converted to a readable format on the frontend.

+++ backend/db/migrations/20210429160804-create-comment.js
@@ -24,7 +24,7 @@ module.exports = {
       },
       timestamp: {
         allowNull: false,
-        type: Sequelize.NUMERIC
+        type: Sequelize.DOUBLE
       },
       text: {
         allowNull: false,
